Cleaned up the way the Raytracer project handles resources.
Previously, the user would have to manually copy the shaders and the
binary resources (dlls, asset databases, etc) into whatever folder the project was
executing from. This now uses pre and post build events to ensure that all
a user has to do is hit run and everything is taken care of for them.

In addition, the output directory and working directory locations have been
cleaned up as well.